finish unit tests, remove raf (no use yet)
