+++ package.json
@@ -21,7 +21,6 @@
     "babel-jest": "^22.4.3",
     "flow-bin": "^0.66.0",
     "jest": "^22.4.3",
-    "raf": "^3.4.0",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
     "react-test-renderer": "^16.3.0",
@@ -34,6 +33,5 @@
     "react-dom": "^16.3.0"
   },
   "description": "React store built on top of the new context API",
-  "repository": "git@github.com:didierfranc/react-waterfall.git",
-  "jest": { "setupFiles": [ "raf/polyfill" ] }
+  "repository": "git@github.com:didierfranc/react-waterfall.git"
 }
